Break out main.rs into bin/flat-manager.rs and lib.rs
This way we can have multiple binaries share code.
